requirements.txt
Django==5.0.2
django-jazzmin==2.6.0
Pillow==10.2.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.30
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
click==3.0.0 
django-modeltranslation==0.18.11
googletrans==4.0.0-rc1
django-leaflet==0.29.0